+++ interp/notation_ops.mli
@@ -89,6 +89,3 @@ val match_notation_constr_ind_pattern :
   inductive -> 'a cases_pattern_g list -> interpretation ->
   (('a cases_pattern_g * extended_subscopes) list * ('a cases_pattern_g list * extended_subscopes) list) *
     (bool * int * 'a cases_pattern_g list)
-
-(** {5 Matching a notation pattern against a [glob_constr]} *)
-